+++ Palace/Book/UI/BookDetail/BookDetailView.swift
@@ -55,7 +55,6 @@ struct BookDetailView: View {
               })
           }
         }
-        .hideFloatingTabBar()
         .edgesIgnoringSafeArea(.all)
         .onChange(of: viewModel.book) { newValue in
           resetSampleToolbar()
@@ -65,11 +64,13 @@ struct BookDetailView: View {
         }
       }
       .onAppear {
+        UITabBarController.hideFloatingTabBar()
         headerHeight = viewModel.isFullSize ? 300 : 225
         viewModel.fetchRelatedBooks()
         self.descriptionText = viewModel.book.summary ?? ""
       }
       .onDisappear {
+        UITabBarController.showFloatingTabBar()
         viewModel.showHalfSheet = false
       }
       .fullScreenCover(item: $selectedBook) { book in
